chore: #91 move txn_signature to file within txn
